- CHANGED: file_manager allways checks if a file existst
git-svn-id: file:///Users/jan/Dropbox/svn/FlexyAdmin/trunk@2405 0c02b2c5-a8bd-4e35-85b2-cf8cacb96951

+++ sys/flexyadmin/models/file_manager.php
@@ -296,10 +296,11 @@ class File_manager Extends CI_Model {
     $result=true;
     
     if (is_dir($name)) {
-      $result=rmdir($name);
+      if (file_exists($name)) $result=rmdir($name);
     }
     else {
-      $result=unlink($name);
+      $result=false;
+      if (file_exists($name)) $result=unlink($name);
 
   		if ($result) {
   			/**
